Added Enitty ApplicationUser, Login,  Initialize from Register user and admin.
Added AccountController
Refactoring HomeController,

Added Link to DB,

In the previous version, the chat works, but there were no authentications and authorizations. At the moment, authentication and authorization have been added, but from the client side we get a login display and a description of the actions that occur when the Admin Account is launched in the webBrowser.